+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file's format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Ensure that derive log level handling still works with `slog` 2.5.
 
 ## [5.2.0]
 ### Changed